Main Build
So learning CSS grids was easier than expected-

While that happened, the js in actually getting it to calculate is off. its probably the number with the parse but ill figure it out tomorrow.

So instead of using 3 weeks for that project, i can probably push it into one at this rate.

-3/4/24